pm_requirements.txt
accelerate==1.8.1
aiofiles==24.1.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
Brotli==1.1.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
diffusers==0.29.1
easydict==1.13
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.116.0
ffmpy==0.6.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.5
fsspec==2025.5.1
gradio==5.36.2
gradio_client==1.10.4
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
insightface==0.7.3
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.1
opencv-python-headless==4.12.0.88
orjson==3.10.18
packaging==25.0
pandas==2.3.1
peft==0.16.0
pillow==11.3.0
prettytable==3.16.0
protobuf==6.31.1
psutil==5.9.8
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruff==0.12.2
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
simsimd==6.5.0
six==1.17.0
sniffio==1.3.1
spaces==0.37.1
starlette==0.46.2
stringzilla==3.12.5
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.1
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
zipp==3.23.0
